Fix the space issue
Brain fart, mental model around shell are messed up.
When shell is invoking the binary it has to parse the cli, so then you
end up with situations where you have to escape separators using quotes.
However go passes the arguments directly to the binary so there is no
such issue. Do not try and escape the arguments, even with spaces.